Разделить JS по файлам
