+++ index.html
@@ -106,7 +106,7 @@
     </div>
   </template>
   <!-- Подключение JavaScript -->
-  <script src="./scripts/script.js"></script>
+  <script type="module" src="./scripts/script.js"></script>
 </body>
 
 </html>
