+++ scripts/constants.js
@@ -1,0 +1,58 @@
+//Шесть карточек «из коробки»
+const initialCards = [
+  {
+    name: 'Архыз',
+    link: 'https://pictures.s3.yandex.net/frontend-developer/cards-compressed/arkhyz.jpg'
+  },
+  {
+    name: 'Челябинская область',
+    link: 'https://pictures.s3.yandex.net/frontend-developer/cards-compressed/chelyabinsk-oblast.jpg'
+  },
+  {
+    name: 'Иваново',
+    link: 'https://pictures.s3.yandex.net/frontend-developer/cards-compressed/ivanovo.jpg'
+  },
+  {
+    name: 'Камчатка',
+    link: 'https://pictures.s3.yandex.net/frontend-developer/cards-compressed/kamchatka.jpg'
+  },
+  {
+    name: 'Холмогорский район',
+    link: 'https://pictures.s3.yandex.net/frontend-developer/cards-compressed/kholmogorsky-rayon.jpg'
+  },
+  {
+    name: 'Байкал',
+    link: 'https://pictures.s3.yandex.net/frontend-developer/cards-compressed/baikal.jpg'
+  }
+];
+
+//Попапы
+const popups = document.querySelectorAll('.popup');
+const popupEdit = document.querySelector('#edit');
+const popupAdd = document.querySelector('#add');
+const addBtn = document.querySelector('.profile__add-btn');
+const editBtn = document.querySelector('.profile__name-edit');
+
+//Попап создания карточки
+const createForm = document.querySelector('#addPlace');
+const createBtn = document.querySelector('#create-btn');
+const photoNameInput = document.querySelector('.popup__element_data_photoname');
+const photoLinkInput = document.querySelector('.popup__element_data_photolink');
+
+//Поля формы "Редактировать профиль"
+const profileName = document.querySelector('.profile__name-text');
+const profileDesc = document.querySelector('.profile__desc');
+const profileInfo = document.querySelector('#profileInfo');
+const nameInput = document.querySelector('.popup__element_data_firstname');
+const jobInput = document.querySelector('.popup__element_data_about');
+
+//Карточки
+const cardTemplate = document.querySelector('#card').content;
+const cards = document.querySelector('.cards');
+const photoFull = document.querySelector('#photo-full');
+const photoFullImage = document.querySelector('.popup__img');
+const photoFullName = document.querySelector('.popup__caption');
+
+//Экспорт
+export {initialCards, popups, popupEdit, popupAdd, addBtn, editBtn, createForm, createBtn, photoNameInput, photoLinkInput, profileName,
+profileDesc, profileInfo, nameInput, jobInput, cardTemplate, cards, photoFull, photoFullImage, photoFullName};
