+++ scripts/script.js
@@ -1,116 +1,10 @@
-//Объявление переменных
-
-//Шесть карточек «из коробки»
-const initialCards = [
-  {
-    name: 'Архыз',
-    link: 'https://pictures.s3.yandex.net/frontend-developer/cards-compressed/arkhyz.jpg'
-  },
-  {
-    name: 'Челябинская область',
-    link: 'https://pictures.s3.yandex.net/frontend-developer/cards-compressed/chelyabinsk-oblast.jpg'
-  },
-  {
-    name: 'Иваново',
-    link: 'https://pictures.s3.yandex.net/frontend-developer/cards-compressed/ivanovo.jpg'
-  },
-  {
-    name: 'Камчатка',
-    link: 'https://pictures.s3.yandex.net/frontend-developer/cards-compressed/kamchatka.jpg'
-  },
-  {
-    name: 'Холмогорский район',
-    link: 'https://pictures.s3.yandex.net/frontend-developer/cards-compressed/kholmogorsky-rayon.jpg'
-  },
-  {
-    name: 'Байкал',
-    link: 'https://pictures.s3.yandex.net/frontend-developer/cards-compressed/baikal.jpg'
-  }
-];
-
-//Попапы
-const popups = document.querySelectorAll('.popup');
-const popupEdit = document.querySelector('#edit');
-const popupAdd = document.querySelector('#add');
-const addBtn = document.querySelector('.profile__add-btn');
-const editBtn = document.querySelector('.profile__name-edit');
-
-//Попап создания карточки
-const createForm = document.querySelector('#addPlace');
-const createBtn = document.querySelector('#create-btn');
-const photoNameInput = document.querySelector('.popup__element_data_photoname');
-const photoLinkInput = document.querySelector('.popup__element_data_photolink');
-
-//Поля формы "Редактировать профиль"
-const profileName = document.querySelector('.profile__name-text');
-const profileDesc = document.querySelector('.profile__desc');
-const profileInfo = document.querySelector('#profileInfo');
-const nameInput = document.querySelector('.popup__element_data_firstname');
-const jobInput = document.querySelector('.popup__element_data_about');
-
-//Карточки
-const cardTemplate = document.querySelector('#card').content;
-const cards = document.querySelector('.cards');
-const photoFull = document.querySelector('#photo-full');
-const photoFullImage = document.querySelector('.popup__img');
-const photoFullName = document.querySelector('.popup__caption');
-
-//Функции
-
-//Открытие попапов
-function openPopup(popupElement) {
-  popupElement.classList.add('popup_opened');
-}
-
-//Закрытие попапов
-function closePopup(popupElement) {
-  const button = popupElement.querySelector('.popup__btn');
-  button.classList.add('popup__btn_inactive');
-  popupElement.classList.remove('popup_opened');
-}
-
-//Редактирование имени и информации о себе
-function editProfile(evt) {
-  evt.preventDefault();
-  profileName.textContent = nameInput.value;
-  profileDesc.textContent = jobInput.value;
-  closePopup(popupEdit);
-}
-
-//Создание карточки
-function createCard(name, link) {
-  const card = cardTemplate.querySelector('.card').cloneNode(true);
-  card.querySelector('.card__name-text').textContent = name;
-  card.querySelector('.card__image').src = link;
-  card.querySelector('.card__image').alt = name;
-  //Удаление карточки
-  const removeBtn = card.querySelector('.card__remove-icon');
-  function removeCard() {
-    card.remove();
-  }
-  removeBtn.addEventListener('click', removeCard);
-  //Превью карточки
-  const cardImage = card.querySelector('.card__image');
-  const cardName = card.querySelector('.card__name-text');
-  cardImage.addEventListener('click', function () {
-    photoFull.classList.add('popup_opened');
-    photoFullImage.src = cardImage.src;
-    photoFullImage.alt = cardName.textContent;
-    photoFullName.textContent = cardName.textContent;
-  });
-  return card;
-}
-
-//Функция, которая добавляет карточку в DOM
-function renderCard(somecard, somecontainer) {
-  somecontainer.prepend(somecard);
-}
-
-//Отрисовка исходных 6 карточек
-initialCards.forEach(function (item) {
-  const myCard = createCard(item.name, item.link);
-  renderCard(myCard, cards);
-})
+//Импорт
+import {initialCards, popups, popupEdit, popupAdd, addBtn, editBtn, createForm, createBtn, photoNameInput, photoLinkInput, profileName,
+profileDesc, profileInfo, nameInput, jobInput, cardTemplate, cards, photoFull, photoFullImage, photoFullName} from './constants.js' ;
+import {showInputError, hideInputError, isValid, hasInvalidInput, toggleButtonState, setEventListeners, enableValidation} from './validate.js';
+import {createCard, renderCard, adddLike} from './card.js';
+import {openPopup, editProfile} from './modal.js';
+import {closePopup} from './utils.js';
 
 //Обработчики
 
@@ -148,96 +42,7 @@ popups.forEach(function (popup) {
 //Редактирование имени и информации о себе
 profileInfo.addEventListener('submit', editProfile);
 
-//Добавление новой карточки
-createForm.addEventListener('submit', function (evt) {
-  evt.preventDefault();
-  renderCard(createCard(photoNameInput.value, photoLinkInput.value), cards);
-  createForm.reset();
-  closePopup(popupAdd);
-})
-
-//Лайк карточки - один обработчик
-function adddLike(evt) {
-  evt.target.classList.toggle('card__like_liked');
-}
-
-cards.addEventListener('click', function (evt) {
-  if (evt.target.classList.contains('card__like')) {
-    adddLike(evt);
-  }
-}
-)
-
 //Валидация форм
-
-//Функции
-//Показать ошибку
-const showInputError = (formElement, inputElement, errorMessage) => {
-  const errorElement = formElement.querySelector(`.${inputElement.id}-error`);
-  inputElement.classList.add('popup__element_type_error');
-  errorElement.textContent = errorMessage;
-  errorElement.classList.add('popup__element-error_active');
-};
-
-//Скрыть ошибку
-const hideInputError = (formElement, inputElement) => {
-  const errorElement = formElement.querySelector(`.${inputElement.id}-error`);
-  inputElement.classList.remove('popup__element_type_error');
-  errorElement.textContent = '';
-  errorElement.classList.remove('popup__element-error_active');
-};
-
-//Валидация поля
-const isValid = (formElement, inputElement) => {
-  if (!inputElement.validity.valid) {
-    showInputError(formElement, inputElement, inputElement.validationMessage);
-  } else {
-    hideInputError(formElement, inputElement);
-  }
-};
-
-//Активность кнопки при валидации
-//Проверка валидности полей
-const hasInvalidInput = (inputList) => {
-  return inputList.some((inputElement) => {
-    return !inputElement.validity.valid;
-  });
-};
-
-//Изменение активности кнопки
-const toggleButtonState = (inputList, buttonElement) => {
-  if (hasInvalidInput(inputList)) {
-    buttonElement.classList.add('popup__btn_inactive');
-  } else {
-    buttonElement.classList.remove('popup__btn_inactive');
-  };
-};
-
-//Обработчики
-//Обработчики полей формы
-const setEventListeners = (formElement) => {
-  const inputList = Array.from(formElement.querySelectorAll('.popup__element'));
-  const buttonElement = formElement.querySelector('.popup__btn');
-  toggleButtonState(inputList, buttonElement);
-  inputList.forEach((inputElement) => {
-    inputElement.addEventListener('input', () => {
-      isValid(formElement, inputElement);
-      toggleButtonState(inputList, buttonElement);
-    });
-  });
-};
-
-//Обработчики форм
-const enableValidation = () => {
-  const formList = Array.from(document.querySelectorAll('.form'));
-  formList.forEach((formElement) => {
-    formElement.addEventListener('submit', (evt) => {
-      evt.preventDefault();
-    });
-    setEventListeners(formElement);
-  });
-};
-
 enableValidation();
