+++ scripts/utils.js
@@ -1,0 +1,7 @@
+//Закрытие попапов
+function closePopup(popupElement) {
+  popupElement.classList.remove('popup_opened');
+}
+
+//Экспорт
+export {closePopup};
